Cast dev and rdev to uint64 in order for openbsd to compile

+++ server.go
@@ -105,8 +105,8 @@ func (s *Server) infoToFileInfo(info os.FileInfo, relativepath, absolutepath str
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
 		fi.Inode = stat.Ino
 		fi.Nlink = uint64(stat.Nlink) // force to uint64 for 32-bit systems
-		fi.Dev = stat.Dev
-		fi.Rdev = stat.Rdev
+		fi.Dev = uint64(stat.Dev)
+		fi.Rdev = uint64(stat.Rdev)
 		fi.Owner = stat.Uid
 		fi.Group = stat.Gid
 		fi.Permissions = stat.Mode
